Yeni Pişti Başlangıç
Started converting the game chkobba to pişti. Game will be in mobile platforms but may also be published in webgl for demo purposes. New rules should be added, or older rules should be discarded to make the game as pişti. First thing to do is to make moves decide on only the top card of the center pile and also make player hands 4 instead of 3