Major Card Idendification Change
Added unique card ID to each card in CardInteraction script so that even if cards are coppied or manupilated to  have different values they can still be spesifically identified. Updated most of the scripts to accommodate the changes made to the cards. Will help with the implementation of some super powers since everything is easier to find now.