Bomba Animation/Restriction and Zafer Puanı Implementation
Added a bomb and explosioın effect for the power Bomba. Also added restrction to the power so that it can only be called if there are cards in the middle. Also added a new power named Zafer puanı that gives extra points if it is still in the players power pool at the end of the round. Had to change how and when the server calls DecideWinner to make sure the Zafer Points reach the server before the desicion is made. Couldnt test the restriction and the new power because Unity Services are Major Incident status