Kapkaç, Oynayamazsın and Kopyala Yapıştır Sync and animtion
Made sure each client synced with the power and added fade in animations for the powers and their indicators. Minor corrections as well. I decided that powers will not reset between the rounds so the cards now actually change value before being sent to the server for kopyala yapıştır and kapkaç. Will add Yak Dünyayı that will make the cards value 0 that works similarly to kapkaç