README for Copilot
Added a readme file that can be used as a reference by Copilot so that copilot's answers can be more consistent and relaiable according to the bigger vision for the game. README icludes the general rules Copilot should follow, to do list and documentation. Now all of these can be given to the Copilot with the promt so that it is never lost and can generate answers that agree with the future of the project.