Major Corrections and Decoupling of GameManager and DeckController
Made a lot of corrections to make sure the game is behaving exatcly how I want. Moved all card transform logic to deck controller and converted to using DOTween instead of lerp functions so that flow can be kept track of more easily. Kutsal Deste and Ver Zehri is changed so that the points are calculated according to the centerCard count when captured. Swap powers were tweeked too and it is a lot better but still needs some work. Also added new sprites for all the powers