Calculator full implementation and Card Selection animation correction
Fully integrated the calculator to the game. Now you can set the value and the number of super power tokens you want to pull and it works correctly (still needs to be tested and corrected for edge cases). Also finally found a way to make sure the card animation sequences are killed correctly before the card is added to the center