Turn and Hand Restrictions and Corrections
Made it so that you can only play your cards and only play them in your turn. There was 2 different input controls both for pc and mobile but turns out the mobile inputs also work for mouse inputs so I was running some of the logic twice at the same time. Also made it so that the infoBox is a gameObject instead of UI.