Menu System BackEnd Integration
Added a script that takes the token prefabs and turns them into items of a menu with names and rarities next to it, so that players have accsess to a menu where they can take a look at the prices of the super powers they want. Logic is integrated but front end changes are necessary too.