Basic GameState Implementation
Now you can save a gameState and later load it to have the views show the saved state on the screen. There are some problems with it though. For example it seems to be not saving the moment the save function was called. Also more test are needed to see if the game can be kept on being played.